+++ University/src/main/java/ua/com/nikiforov/models/persons/Teacher.java
@@ -2,14 +2,14 @@ package ua.com.nikiforov.models.persons;
 
 public class Teacher extends Person {
 
-    private long subjectId;
+    private int subjectId;
 
     public Teacher(long id, String firstName, String lastName, int subjectId) {
         super(id, firstName, lastName);
         this.subjectId = subjectId;
     }
 
-    public long getSubjectId() {
+    public int getSubjectId() {
         return subjectId;
     }
